remove red zone usage below the stack
the red zone is a 128 byte area below the stack pointer that functions
can use for temporary storage without adjusting %rsp. this becomes an
issue when a function is using this space, an interrupt is triggered,
and the cpu pushes data onto the stack, potentially overwriting the
previously used data. it's a good idea to have this disabled in kernel
code.