implement log abstraction and use log -> serial for x86_64
the log prints in the apic code that came from i386 included a
dependency with the VGA text mode via the tty driver. instead, we moved
to an abstraction around the log functions and pipe into serial_write
for x86_64. this allows us to mostly keep the same code, as well as
improving i386 by using proper log functions insstead of just printf's