move from Makefile to cmake
there is a desire to migrate to uAPIC for the x86_64 architecture,
meaning we have to switch from using Makefiles to cmake. This is a
welcome move however, since the previous bash scripts + nested
independent Makefiles made for a very hacky setup. whoever made the
skeleton tutorials in osdev did not like their readers.